dashboardUI-2, comps added
