+++ src/app/svg/FilesSVG/Image.tsx
@@ -1,0 +1,40 @@
+const ImageIcon = () => {
+    return (
+        <div>
+            <svg
+                viewBox="0 0 24 24"
+                width="120px"
+                height="120px"
+                fill="none"
+                xmlns="http://www.w3.org/2000/svg"
+            >
+                <g id="SVGRepo_bgCarrier" strokeWidth="0"></g>
+                <g
+                    id="SVGRepo_tracerCarrier"
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                ></g>
+                <g id="SVGRepo_iconCarrier">
+                    {" "}
+                    <path
+                        opacity="0.15"
+                        d="M4 5C4 4.44772 4.44772 4 5 4H19C19.5523 4 20 4.44772 20 5V19C20 19.5523 19.5523 20 19 20H5C4.44772 20 4 19.5523 4 19V5Z"
+                        // fill="#000000"
+                        fill="#1C274C"
+                    ></path>{" "}
+                    <path
+                        d="M4.02693 18.329C4.18385 19.277 5.0075 20 6 20H18C19.1046 20 20 19.1046 20 18V14.1901M4.02693 18.329C4.00922 18.222 4 18.1121 4 18V6C4 4.89543 4.89543 4 6 4H18C19.1046 4 20 4.89543 20 6V14.1901M4.02693 18.329L7.84762 14.5083C8.52765 13.9133 9.52219 13.8481 10.274 14.3494L10.7832 14.6888C11.5078 15.1719 12.4619 15.1305 13.142 14.5864L15.7901 12.4679C16.4651 11.9279 17.4053 11.8855 18.1228 12.3484C18.2023 12.3997 18.2731 12.4632 18.34 12.5301L20 14.1901M11 9C11 10.1046 10.1046 11 9 11C7.89543 11 7 10.1046 7 9C7 7.89543 7.89543 7 9 7C10.1046 7 11 7.89543 11 9Z"
+                        stroke="#0a0a0a"
+                        fill="#1C274C"
+                        opacity="0.5"
+                        strokeWidth="1.2"
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                    ></path>{" "}
+                </g>
+            </svg>
+        </div>
+    );
+};
+
+export default ImageIcon;
